test(node): add task to run a subset of tests on node
`npm run test-node` or `gulp test/node` will now run tests
on node.js. These are also run on TravisCI. Clean up gulp
build.
