+++ .travis.yml
@@ -24,4 +24,4 @@ before_script:
 
 script:
   - node_modules/.bin/karma start karma-sauce.conf.js
-
+  - node_modules/.bin/gulp test/node
