--- source_map_test.html
@@ -1,3 +1,0 @@
-<html>
-  <script src="build/test/source_map_test.js"></script>
-</html>
